requirements.txt
llama_index=0.9.43
numpy==1.24.4
openai==1.3.8
openai==1.3.8
pandas==2.1.1
PyPDF2==3.0.1
spacy==3.7.1
streamlit==1.31.0
qdrant-client==1.7.2
tabula-py==2.9.0
transformers==2.2.2